requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
dnspython==1.16.0
gunicorn==19.9.0
honcho==1.0.1
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
